- refactor : 단일 회원 찾기 서비스 패키지 명 수정
