+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/member/FindIdPwdController.java
@@ -1,7 +1,7 @@
 package ProjectDoge.StudentSoup.controller.member;
 
 import ProjectDoge.StudentSoup.dto.member.EmailDto;
-import ProjectDoge.StudentSoup.service.member.MemberFindService;
+import ProjectDoge.StudentSoup.service.member.MemberAccountFindService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,19 +17,19 @@ import java.util.Map;
 @RestController
 public class FindIdPwdController {
 
-    private final MemberFindService memberFindService;
+    private final MemberAccountFindService memberAccountFindService;
 
     @PostMapping("/find/id")
     public String findAccountMemberId(@RequestBody Map<String, String> map){
-        EmailDto emailDto = memberFindService.createFindMemberIdUsingEmail(map.get("email"));
-        memberFindService.mailSend(emailDto);
+        EmailDto emailDto = memberAccountFindService.createFindMemberIdUsingEmail(map.get("email"));
+        memberAccountFindService.mailSend(emailDto);
         return "ok";
     }
 
     @PostMapping("/find/pwd")
     public String findAccountMemberPwd(@RequestBody Map<String, String> map){
-        EmailDto emailDto = memberFindService.createFindPwdUsingEmailAndId(map.get("email"), map.get("id"));
-        memberFindService.mailSend(emailDto);
+        EmailDto emailDto = memberAccountFindService.createFindPwdUsingEmailAndId(map.get("email"), map.get("id"));
+        memberAccountFindService.mailSend(emailDto);
         return "ok";
     }
 }
